Refactor to be just one assist
