+++ crates/ide_assists/src/lib.rs
@@ -186,7 +186,6 @@ mod handlers {
     mod convert_iter_for_each_to_for;
     mod convert_into_to_from;
     mod convert_tuple_struct_to_named_struct;
-    mod convert_tuple_variant_to_named_variant;
     mod early_return;
     mod expand_glob_import;
     mod extract_function;
@@ -257,7 +256,6 @@ mod handlers {
             convert_iter_for_each_to_for::convert_iter_for_each_to_for,
             convert_into_to_from::convert_into_to_from,
             convert_tuple_struct_to_named_struct::convert_tuple_struct_to_named_struct,
-            convert_tuple_variant_to_named_variant::convert_tuple_variant_to_named_variant,
             early_return::convert_to_guarded_return,
             expand_glob_import::expand_glob_import,
             extract_struct_from_enum_variant::extract_struct_from_enum_variant,
